enabling cache flag
