+++ README.md
@@ -47,7 +47,7 @@ let ode = simulator::Equation::new_ode(
     (2, 1),
 );
 
-let op = ode.simulate_subject(&subject, &vec![0.3, 0.5, 0.1, 70.0]);
+let op = ode.simulate_subject(&subject, &vec![0.3, 0.5, 0.1, 70.0], false);
 println!("{op:#?}");
 ```
