+++ examples/ode_readme.rs
@@ -39,6 +39,6 @@ fn main() {
         (2, 1),
     );
 
-    let op = ode.simulate_subject(&subject, &vec![0.3, 0.5, 0.1, 70.0]);
+    let op = ode.simulate_subject(&subject, &vec![0.3, 0.5, 0.1, 70.0], false);
     println!("{op:#?}");
 }
