+++ src/lib.rs
@@ -16,8 +16,10 @@ pub mod prelude {
     }
     pub mod simulator {
         pub use crate::simulator::{
-            get_population_predictions,
-            likelihood::{pf_psi, PopulationPredictions, Prediction, SubjectPredictions},
+            likelihood::{
+                get_population_predictions, pf_psi, PopulationPredictions, Prediction,
+                SubjectPredictions,
+            },
             Equation,
         };
     }
